+++ backend/src/app/entities/apphttpexception.entity.ts
